[REFACTOR][CORE] Move file checksum logic to separate package
Moves the logic dealing with filesystem checksum caches into the
separate subpackage 'filesystem.checksum'. This makes it easier to
differentiate which classes belong to the filesystem implementation.
