+++ core/src/saros/filesystem/checksum/IFileContentChangedListener.java
@@ -1,4 +1,6 @@
-package saros.filesystem;
+package saros.filesystem.checksum;
+
+import saros.filesystem.IFile;
 
 public interface IFileContentChangedListener {
   public void fileContentChanged(IFile file);
