+++ core/src/saros/filesystem/checksum/IFileContentChangedNotifier.java
@@ -1,4 +1,4 @@
-package saros.filesystem;
+package saros.filesystem.checksum;
 
 public interface IFileContentChangedNotifier {
