+++ core/src/saros/negotiation/AbstractIncomingProjectNegotiation.java
@@ -16,11 +16,11 @@ import saros.communication.extensions.StartActivityQueuingResponse;
 import saros.exceptions.LocalCancellationException;
 import saros.exceptions.SarosCancellationException;
 import saros.filesystem.FileSystem;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IFolder;
 import saros.filesystem.IProject;
 import saros.filesystem.IResource;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.IProgressMonitor;
 import saros.monitoring.SubProgressMonitor;
 import saros.negotiation.NegotiationTools.CancelOption;
