+++ core/src/saros/negotiation/AbstractOutgoingProjectNegotiation.java
@@ -15,9 +15,9 @@ import saros.communication.extensions.StartActivityQueuingResponse;
 import saros.editor.IEditorManager;
 import saros.exceptions.LocalCancellationException;
 import saros.exceptions.SarosCancellationException;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IProject;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.IProgressMonitor;
 import saros.monitoring.SubProgressMonitor;
 import saros.negotiation.NegotiationTools.CancelOption;
