+++ core/src/saros/negotiation/FileListFactory.java
@@ -8,11 +8,11 @@ import java.util.LinkedList;
 import java.util.List;
 import org.apache.log4j.Logger;
 import saros.filesystem.FileSystem;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IFile;
 import saros.filesystem.IFolder;
 import saros.filesystem.IProject;
 import saros.filesystem.IResource;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.IProgressMonitor;
 import saros.monitoring.NullProgressMonitor;
 import saros.negotiation.FileList.MetaData;
