+++ core/src/saros/negotiation/InstantIncomingProjectNegotiation.java
@@ -9,9 +9,9 @@ import org.apache.commons.io.input.CountingInputStream;
 import org.apache.log4j.Logger;
 import saros.exceptions.LocalCancellationException;
 import saros.exceptions.SarosCancellationException;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IProject;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.IProgressMonitor;
 import saros.negotiation.NegotiationTools.CancelOption;
 import saros.negotiation.stream.IncomingStreamProtocol;
