+++ core/src/saros/negotiation/InstantOutgoingProjectNegotiation.java
@@ -19,10 +19,10 @@ import saros.editor.ISharedEditorListener;
 import saros.editor.remote.UserEditorStateManager;
 import saros.exceptions.LocalCancellationException;
 import saros.exceptions.SarosCancellationException;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IFile;
 import saros.filesystem.IProject;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.IProgressMonitor;
 import saros.negotiation.NegotiationTools.CancelOption;
 import saros.negotiation.stream.OutgoingStreamProtocol;
