+++ core/src/saros/negotiation/NegotiationFactory.java
@@ -3,8 +3,8 @@ package saros.negotiation;
 import java.util.List;
 import saros.context.IContainerContext;
 import saros.editor.IEditorManager;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.negotiation.hooks.SessionNegotiationHookManager;
 import saros.net.IConnectionManager;
 import saros.net.IReceiver;
