+++ core/src/saros/negotiation/ProjectNegotiation.java
@@ -8,8 +8,8 @@ import saros.communication.extensions.CancelProjectNegotiationExtension;
 import saros.exceptions.LocalCancellationException;
 import saros.exceptions.RemoteCancellationException;
 import saros.exceptions.SarosCancellationException;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.IProgressMonitor;
 import saros.monitoring.MonitorableFileTransfer;
 import saros.monitoring.MonitorableFileTransfer.TransferStatus;
