+++ core/test/junit/saros/SarosCoreContextFactoryTest.java
@@ -9,8 +9,8 @@ import saros.context.IContainerContext;
 import saros.context.IContextFactory;
 import saros.context.IContextKeyBindings;
 import saros.editor.IEditorManager;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.remote.IRemoteProgressIndicatorFactory;
 import saros.preferences.IPreferenceStore;
 import saros.preferences.Preferences;
