+++ eclipse/src/saros/SarosEclipseContextFactory.java
@@ -14,11 +14,11 @@ import saros.editor.EditorManager;
 import saros.editor.IEditorManager;
 import saros.filesystem.EclipsePathFactory;
 import saros.filesystem.EclipseWorkspaceImpl;
-import saros.filesystem.FileContentNotifierBridge;
-import saros.filesystem.FileSystemChecksumCache;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IPathFactory;
 import saros.filesystem.IWorkspace;
+import saros.filesystem.checksum.FileContentNotifierBridge;
+import saros.filesystem.checksum.FileSystemChecksumCache;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.remote.EclipseRemoteProgressIndicatorFactoryImpl;
 import saros.monitoring.remote.IRemoteProgressIndicatorFactory;
 import saros.preferences.EclipsePreferenceStoreAdapter;
