+++ eclipse/src/saros/ui/wizards/AddProjectToSessionWizard.java
@@ -44,8 +44,8 @@ import org.eclipse.ui.IFileEditorInput;
 import saros.Saros;
 import saros.SarosPluginContext;
 import saros.editor.internal.EditorAPI;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.ResourceAdapterFactory;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.monitoring.ProgressMonitorAdapterFactory;
 import saros.negotiation.AbstractIncomingProjectNegotiation;
 import saros.negotiation.CancelListener;
