+++ intellij/src/saros/intellij/context/SarosIntellijContextFactory.java
@@ -8,10 +8,10 @@ import saros.core.ui.eventhandler.UserStatusChangeHandler;
 import saros.core.ui.eventhandler.XMPPAuthorizationHandler;
 import saros.core.util.IntellijCollaborationUtilsImpl;
 import saros.editor.IEditorManager;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IPathFactory;
 import saros.filesystem.IWorkspace;
-import saros.filesystem.NullChecksumCache;
+import saros.filesystem.checksum.IChecksumCache;
+import saros.filesystem.checksum.NullChecksumCache;
 import saros.intellij.editor.EditorManager;
 import saros.intellij.negotiation.ModuleConfigurationProvider;
 import saros.intellij.negotiation.hooks.ModuleTypeNegotiationHook;
