+++ intellij/src/saros/intellij/ui/wizards/AddProjectToSessionWizard.java
@@ -29,8 +29,8 @@ import org.apache.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import saros.SarosPluginContext;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IProject;
+import saros.filesystem.checksum.IChecksumCache;
 import saros.intellij.context.SharedIDEContext;
 import saros.intellij.editor.DocumentAPI;
 import saros.intellij.filesystem.IntelliJProjectImpl;
