+++ server/src/saros/server/ServerContextFactory.java
@@ -7,11 +7,11 @@ import org.apache.commons.io.FileUtils;
 import org.apache.log4j.Logger;
 import saros.context.AbstractContextFactory;
 import saros.context.IContextKeyBindings;
-import saros.filesystem.IChecksumCache;
 import saros.filesystem.IPath;
 import saros.filesystem.IPathFactory;
 import saros.filesystem.IWorkspace;
-import saros.filesystem.NullChecksumCache;
+import saros.filesystem.checksum.IChecksumCache;
+import saros.filesystem.checksum.NullChecksumCache;
 import saros.monitoring.remote.IRemoteProgressIndicatorFactory;
 import saros.preferences.IPreferenceStore;
 import saros.preferences.Preferences;
